# projetoConclusaoPos/vacine/vacine-backend/api-swagger.yaml
openapi: 3.0.3
info:
  title: API para Clínica de Imunizacação - Vacine
  description: |-
    
    API para clínica de imunização Vacine.
    
     Faz autenticação, autorização e registro de usuários. Além disso, ainda possibilita a gestão de usuários, vacinas, fornecedores, vacinacões, compras e descartes de vacinas. 
    
    Esta API foi criada via Swagger com base na especificação OpenAPI 3.0. 

    Instruções:
     - Para acessar as funções pelo servidor 'API Servidor' ou 'API Desenvolvimento':
       1. Acessar a API de 'Autorização'.
       2. Utilizar o endpoint '/autenticacao/jwt' informando um usuário e senha válidos.
       3. Copiar o token da resposta.
       4. Clicar na função 'Authorize'.
       5. Informar o token recebido no passo 3.    
     - Para acessar as funções pelo servidor 'API Mock':
       1. Clicar na função 'Authorize'.
       2. Informar qualquer valor no campo 'value' para o token.

       
    Links úteis:
    - [Repositório do projeto](https://github.com/georgiamcs/posPucMinas2022)
    

  contact:
    name: desenvolvedor
    email: georgiamcs@gmail.com
  version: 1.0.0
...
servers:
  - url: https://georgia2344.c42.integrator.host/mock/api
    description: API Mock
  - url: https://georgia2344.c42.integrator.host/api
    description: API Servidor    
  - url: http://localhost:3005/api
    description: API Desenvolvimento  
...
paths:
...
  /usuarios/trocarsenha/{id}:
    parameters:
      - name: id
        in: path
        description: Id do usuário
        required: true
        schema:
          type: string  
    put:
      tags:
        - Clientes
      summary: Altera a senha do usuário
      operationId: trocarSenha
      requestBody:
        required: true
        description: Senha alterada
        content:
          application/json:
            schema:
                properties:
                  senha:
                    type: string
                    description: Senha alterada do usuário
                    example: "georgiamcs@gmail.com"                  
                required:
                  - senha   
      responses:
        '200':
          description: Operação realizada com sucesso
        '400':
          description: Id inválido.            
        '401':
          description: Acesso não autorizado. Token de autorização não informado ou inválido.       
        '403':
          description: Sem permissão para acessar o serviço
        '404':
          description: Registro não localizado       
        '500':
          description: Erro no processamento da requisição no servidor.              
      security:
        - api_key:
            - write:usuarios
            - read:usuarios
...
  /usuarios/vacinacoes/{id}:
    parameters:
      - name: id
        in: path
        description: Id do usuário
        required: true
        schema:
          type: string  
    get:
      tags:
        - Usuários
      summary: Retorna a lista de vacinacoes de um usuário por id
      operationId: getVacinacoesUsuario
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/Vacinacao'            
        '400':
          description: Id inválido            
        '401':
          description: Acesso não autorizado. Token de autorização não informado ou inválido       
        '403':
          description: Sem permissão para acessar o serviço
        '404':
          description: Registro não localizado          
        '500':
          description: Erro no processamento da requisição no servidor 
      security:
        - api_key:
            - write:usuarios
            - read:usuarios                                
...
components:
  schemas:
...
    UsuarioRelacionamento:
      required:
        - _id
        - nome
      type: object
      properties:
        _id:
          type: string
          description: Id do usuário
          example: "37dc9c85fd018a07bd7eec24"
        nome:
          type: string
          description: Nome do usuário
          example: "Geórgia Maria Carvalho de Sousa Galvão"
        cpf:
          type: string
          description: CPF do usuário
          example: "12345678911" 
...
    Vacinacao: 
      required:
        - _id
        - codigo
        - usuario_resp_cadastro
        - usuario_cliente
        - usuario_aplicador_vacina
        - data_aplicacao
        - vl_total
        - dt_inclusao
        - itens_vacinacao
        - __v  
      type: object
      properties:
        _id:
          type: string
          description: Id da vacinação
          example: "37dc9c85fd018a07bd7eec24"
          readOnly: true
        codigo:
          type: string
          description: Código da vacinação
          example: "V2345"
        usuario_resp_cadastro:
          $ref: '#/components/schemas/UsuarioRelacionamento'   
        usuario_cliente:
          $ref: '#/components/schemas/UsuarioRelacionamento' 
        usuario_aplicador_vacina:
          $ref: '#/components/schemas/UsuarioRelacionamento'  
        data_aplicacao:
          type: string
          format: date-time
          description: Data da aplicação da vacina
        vl_total: 
          type: 'number'
          description: Valor total cobrado pela vacinação
          example: 1410.01
        itens_vacinacao: 
          type: 'array'
          items: 
            $ref: '#/components/schemas/ItemVacinacao' 
        dt_inclusao:
          type: string
          format: date-time
          description: Data da inclusão do registro
          readOnly: true  
        dt_alteracao:
          type: string
          format: date-time
          description: Data da alteração do registro
          readOnly: true  
        __v:
          type: integer
          description: Código de controle de versão do banco de dados
          example: 0
          readOnly: true     
    VacinaRelacionamento:
      required:
        - _id
        - nome
        - protecao_contra
      type: object          
      properties:
        _id:
          type: string
          description: Id da vacina
          example: "6400091a4390d93c928c494e"
        nome:
          type: string
          description: Nome da vacina
          example: "Difteria e Tétano (dT)"
        protecao_contra:
          type: string
          description: Descrição de para quais doenças a vacina protege
          example: "Difteria e Tétano" 
    ItemVacinacao:
      required:   
        - vacina
        - lote
        - data_validade
        - qtd_doses
        - vl_item
      type: object          
      properties:
        vacina: 
          $ref: '#/components/schemas/VacinaRelacionamento' 
        lote:
          type: string
          description: Lote da vacina
          example: "L2365" 
        data_validade:
          type: string
          format: date-time
          description: Data de validade da vacina
        qtd_doses:
          type: number
          description: Quantidade de doses aplicadas
          example: 1   
        vl_item:
          type: number
          description: Valor da vacina aplicada
          example: 100           
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer    
      bearerFormat: JWT